Move the checks for conflicting presentations to a standalone action class
