+++ app/Livewire/Schedule/Cell.php
@@ -2,6 +2,7 @@
 
 namespace App\Livewire\Schedule;
 
+use App\Actions\Schedule\PresentationConflictChecker;
 use App\Models\Presentation;
 use App\Models\Room;
 use App\Models\Timeslot;
@@ -21,6 +22,7 @@ class Cell extends Component
         $this->timeslot = $timeslot;
         $this->room = $room;
         $this->id = "r-{$this->room->id}-t-{$this->timeslot->id}";
+
         $this->presentations = $this->room->presentations()
             ->where('start', '>=', Carbon::parse($this->timeslot->start))
             ->where('start', '<', Carbon::parse($this->timeslot->start)->copy()->addMinutes(30))
@@ -35,86 +37,15 @@ class Cell extends Component
             $room = Room::find($newRoom);
             $timeslot = Timeslot::find($newTimeslot);
 
-            if ($this->isClearOfConflicts($presentation, $room, $timeslot)) {
-                $this->dispatchMovingEventToGrid($presentation->id, $room->id, $timeslot->id, $timeslot->start);
+            $presentationChecker = new PresentationConflictChecker();
+            if ($presentationChecker->isClearOfConflicts($presentation, $room, $timeslot->start)) {
+                $this->dispatchMoveEventToGrid($presentation->id, $room->id, $timeslot->id, $timeslot->start);
             }
         }
     }
 
-    public function isClearOfConflicts($presentation, $room, $timeslot)
-    {
-        $previousPresentation = $this->findConflictPresentationBefore($room, $timeslot->start, $presentation);
-        $nextPresentation = $this->findConflictPresentationAfter($room, $timeslot->start, $presentation);
 
-        // Case 1: No presentations in the room before this one or after it that causes conflict
-        if (is_null($previousPresentation)
-            && is_null($nextPresentation)) {
-            return true;
-        }
-
-        // Case 2: Next presentation that causes conflict is actually the same presentation trying to be moved later
-        if (!is_null($nextPresentation) && $presentation->id == $nextPresentation->id && is_null($previousPresentation)) {
-            return true;
-        }
-
-        // Case 3: Previous presentation that causes conflict is
-        // actually the same presentation trying to be moved earlier
-        if (!is_null($previousPresentation) && $presentation->id == $previousPresentation->id && is_null($nextPresentation)) {
-            return true;
-        }
-
-        return false;
-    }
-
-    public function findConflictPresentationBefore($room, $suggestedStart, $presentation)
-    {
-        $presentationStartTime = Carbon::parse($suggestedStart);
-
-        $potentialConflict = $room->presentations()
-            ->where('start', '<', $presentationStartTime)
-            ->orderBy('start', 'desc')
-            ->first();
-
-        if (is_null($potentialConflict)) {
-            return null;
-        }
-
-        $duration = $potentialConflict->type == 'workshop' ?
-            Presentation::$WORKSHOP_DURATION :
-            Presentation::$LECTURE_DURATION;
-
-        return Carbon::parse($potentialConflict->start)->copy()->addMinutes($duration) <= $presentationStartTime
-            ? null
-            : $potentialConflict;
-    }
-
-    public function findConflictPresentationAfter($room, $suggestedStart, $presentation)
-    {
-        $duration = $presentation->type == 'workshop' ?
-            Presentation::$WORKSHOP_DURATION :
-            Presentation::$LECTURE_DURATION;
-        $presentationStartTime = Carbon::parse($suggestedStart);
-        $presentationEndTime = $presentationStartTime->copy()->addMinutes($duration);
-
-        $potentialConflict = $room->presentations()
-            ->where('start', '>=', $presentationStartTime)
-            ->orderBy('start', 'asc')
-            ->first();
-
-        if (is_null($potentialConflict)) {
-            return null;
-        }
-
-        $duration = $potentialConflict->type == 'workshop' ?
-            Presentation::$WORKSHOP_DURATION :
-            Presentation::$LECTURE_DURATION;
-
-        return $presentationEndTime <= Carbon::parse($potentialConflict->start)
-            ? null
-            : $potentialConflict;
-    }
-
-    public function dispatchMovingEventToGrid($id, $newRoom, $newTimeslot, $newTime)
+    public function dispatchMoveEventToGrid($id, $newRoom, $newTimeslot, $newTime)
     {
         $this->dispatch('move-presentation', data: [
             'presentation_id' => $id,
